+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.6.0 2024-04-29
+## 0.7.0 2024-04-04
+
+### Changed
+
+- The `data_type` field of the `BaseField` entity is now an enum of acceptable types.
+- The `ProposalFieldValue` entity now has a `is_valid` field.
+
+## 0.6.0 2024-03-29
 
 ### Changed
 
@@ -17,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The `/applicationForms/:id` endpoint no longer has a `includeFields` query parameter, as it always includes deep fields.
 
-## 0.5.0 - 2024-04-27
+## 0.5.0 - 2024-03-27
 
 ### Added